Learner AI Code Refactoring and removed unnecessary processing
