+++ src/mongodb/config/language/common/commonConfig.ts
@@ -1,0 +1,15 @@
+var lang_common_config = {
+    fluencyCalPerc: {
+        wer: 5,
+        cercal: 10,
+        charCount: 10,
+        wordCount: 10,
+        repetitions: 10,
+        pauseCount: 10,
+        ins: 20,
+        del: 15,
+        sub: 5
+    }
+};
+
+export default lang_common_config;
