+++ src/mongodb/config/language/ta.ts
@@ -1,0 +1,6 @@
+var ta_config = {
+    vowel: ['ா', 'ி', 'ீ', 'ு', 'ூ', 'ெ', 'ே', 'ை', 'ொ', 'ோ', 'ௌ', '்',],
+    language_code: "ta"
+};
+
+export default ta_config;
